Create the `litebox_common_linux` crate, starting with `errno` (#34)
This PR creates the `litebox_common_linux` crate (see #22), which will
house bits that are common across platforms/shims on Linux systems. This
PR only focuses on error numbers for now, since those are a large chunk
of duplication across two existing crates.

This is not just a simple moving of code, but a refactor of the design
of `Errno` to make it both more ergonomic and safe to use, compared to
the `i32` based design used earlier.

The generated functionality (both for associated constants, as well as
`as_str` for better human-readability for display/debugging purposes) in
`generated.rs` relies on `/usr/bin/errno` and custom `awk` scripts that
are included as doc strings in `generated.rs`. I don't expect us to need
to re-generate these often (or even at all), but they exist there
nonetheless.