Update semver-checks workflow to provide stdout feedback (#40)
It got a little annoying to look at the semver-checks changes, so this PR updates it to place it as a comment, which should hopefully be easier to see. When placing the comment, it removes older comments it has made, so that we are not flooded with such comments, but it acts as an easy look-up for the changes.

This PR also disables the reaction thing, which (while useful) was not sufficiently discoverable, and thus is not ideal.