Add an integration test for runner (#155)
For dynamically linked executable, we need to prepare a rootfs tar file
containing all the dependencies it needs. This PR adds an integration
test for the runner that compiles a simple C program, creating a rootfs,
and runs it w/ seccomp (w/o rewriter).

Currently, there are still some syscalls we don't support but the rewriter
would intercept all syscalls and route them to litebox. With seccomp, we
ony selectively intercept syscall we supported.

The test also outputs the command so we can run it manually (for testing
purposes).