Support clone3/set_tid_address/gettid (#158)
This PR adds a new thread provider for spawning new thread and terminate
the calling thread.

Since the new thread does not share the same stack as the calling
thread, we need to divert its execution to some trampoline code that
could restore its registers to resume the execution. Instead of just
passing the syscall arguments, now it passes the execution context
(i.e., `pt_regs` used in Linux kernel).

To make the interface general enough, certain actions that supposed to
be taken by Linux kernel now is done in the shim layer. In particular,

```
CLONE_CHILD_CLEARTID: Clear  (zero)  the  child  thread ID at the location pointed to by child_tid (clone()) or cl_args.child_tid (clone3()) in child memory when the child exits, and do a wakeup on the futex at that address.  The address involved may be changed by the set_tid_address(2) system call.  This is used by threading libraries.
```
This is now done in `sys_exit` in the shim layer. One minor concern is
that if the thread is forcefully killed, we don't have the chance to do
the cleanup. In contrast, kernel always does it.

Since `set_tid_address` can change it, I also add the support for this
syscall, which in turn requires the support for `gettid`.

```
CLONE_CHILD_SETTID (since Linux 2.5.49)
              Store the child thread ID at the location pointed to by child_tid (clone()) or cl_args.child_tid (clone3()) in the child's memory.  The store operation completes before the clone call returns  control  to  user space in the child process.  (Note that the store operation may not have completed before the clone call returns in the parent process, which will be relevant if the CLONE_VM flag is also employed.)

CLONE_PARENT_SETTID (since Linux 2.5.49)
              Store the child thread ID at the location pointed to by parent_tid (clone()) or cl_args.parent_tid (clone3()) in the parent's memory.  (In Linux 2.5.32-2.5.48 there was a flag CLONE_SETTID that did
              this.)  The store operation completes before the clone call returns control to user space.
```
These two can be trivially done in the shim layer after the new thread
is spawned.