LVBS kernel module signature verification (#182)
This PR adds support for verifying Linux kernel module signature. If
module signing is enabled, a Linux kernel build pipeline signs each
kernel module and appends the signature to it. We extract and verify the
signature of each kernel module before validating its relocations and
changing its page permissions.
This signature verification is based on the system certificate which
contains the public portion of the build pipeline key. It is provided by
the VTL0 kernel and locked down before the `end_of_boot` signal. Its
authenticity depends on UEFI Secure Boot which ensures only trusted
software is loaded during the boot process. Also, both the VTL0 kernel
and VTL1 do not have access to the private portion of the key.