Reserve space for trampoline code (#198)
As mentioned in #191, the trampoline code must be loaded right after the
executable code, so our elf loader tries to `mmap` more memory at the
beginning. However, for those libs loaded by `ld`, it is not aware of
trampoline code and hence there might be no space for loading
trampoline.

Instead of fixing it in the loader, this PR takes a different approach,
which modifies `mmap` to make sure that whenever we `mmap` a file with
zero file offset and no MAP_FIXED flag, there must be enough memory
space (16MB) after the mapping. Note that this is done when searching
for available memory regions, it does not actually reserve the 16MB
space.

This PR also fixes the interface for `create_mapping` where `PageRange`
is decoupled into `Option<NonZeroAddress>` and `NonZeroPageSize`.