Migrate and fix design of events interfaces into LiteBox crate (#225)
Our existing interface was a clean-up/generalization of the one in
VSBox. However, we were not perfectly happy with it. Thus, @CvvT
(thanks!) recognized this and implemented a more manageable design into
the Linux shim. However, this meant that we had multiple different
interfaces for events, with differing functionalities and interfaces.

This PR _effectively_ is a migration of (a version of) that into the
main LiteBox crate. Specifically, it removes the old event-manager
design, and instead sets things up purely using observer and polling
interfaces.

Along the way, I have not only added documentation, but also cleaned up
the interfaces to have more easy-to-use ergonomics and functionality.
Additionally, there was a performance bug that existed in the shim that
(in the process of making sure there is a good API specification) that
was discovered---this has been fixed too.

I think that the design can still be improved a bit more, but is good
enough for now, and already cleans some things up to the point that I
would be happy enough to have it be part of the explicit public
interface of LiteBox.

---

Note: This PR also unblocks the _next_ step towards switching the
descriptor interfaces to being fully within LiteBox (as part of the
series begun by https://github.com/microsoft/litebox/pull/180).