Support for OP-TEE TA command loop and queues (#263)
This PR adds support for OP-TEE command loop and queues to enable TA's
event-driven execution. In particular, it executes a TA within a loop while
consuming commands from the submission queue and storing outcomes in
the completion queue. Input and output exchange is performed through
a TA-specific stack. All input commands to a TA can be written in a JSON
file which acts as a client app (mainly for testing and debugging). This
PR contains two example TAs (hello world and RNG) and JSON files to
interact with them.