Clean up RawMutex unblock status logic + Fix mock RawMutex (#279)
This PR clarifies the requirements of the platforms regarding when a
wakeup can occur on a blocked thread. Specifically, it de-restricts the
scenario to be one where a thread can be unblocked even if there has not
been a change to the underlying value, as long as a wake is triggered.

These semantics remove a conditional that existed in each of the
platforms, as well as help de-restrict the upcoming design of the new
futex layer.

There was also a bug in the mock platform's RawMutex support that would
sometimes lead to missed wake-ups, which led to incredibly bizarre and
hard-to-debug behaviors. This PR fixes that too.

Also, as a drive-by, this PR also removes a comment that no longer
applies.