Update exit handling + Multiple clippy fixups for Windows (#315)
This PR is largely focused around clippy fixups, but along the way (due
to the clippy warnings that showed up on Windows), I noticed that our
Linux shim exit usage was accidentally assuming that it knows the actual
underlying type of the `ExitCode`. Thus, I've also fixed that up in this
PR. Well, partly---this actually adds a TODO on how we actually want to
expose the exit code to the outside world; the `ExitProvider` design
explicitly does not support arbitrary exit codes, since there are
platforms where this might even exist, but we then need to think about
how we handle this, or if it even matters, if we are assuming
single-process world). Either way, that TODO should allow us to look
into this later.