LayeredFS: Managing upper layer directories when lower ones exist (#326)
The Layered FS must keep a consistent view between the layers it needs
to deal with, which is complicated at times due to directories. This is
why, for example, we already have a notion of migrating the directories
to the upper layer whenever we perform a `mkdir` command. Basically,
`mkdir a/b` should not fail just because the upper layer does not have
`a/`; instead, it looks up the lower layer, and "copies" the directory
structure over (it does not copy the files over, it just maintains the
minimal upper-layer data to handle this).

I had already implemented this migration for `mkdir`, but had left in a
`TODO` for actual file migrations. Specifically, if `bar/baz` exists at
the lower layer (but there is no `bar/` in the upper layer) and you
attempt to write to it, the file must migrate from lower to upper layer
before it can be written to. However, this requires that `bar/` exists
at the upper layer.

This PR does exactly this sort of syncing. Specifically, it refactors
the `mkdir` code to factor out the common "migrate directories up"
functionality, and then starts using this in the right spots to allow
this to naturally occur during file writes. There is also similar (but
mildly different) behavior needed during file creation.

---

As a side-note: there was a subtle bug in the old `write` behavior where
in some cases, writes to a migrated file might not persist after the
file is closed, due to how we handle the garbage-collection necessary
for files. I've added the necessary `drop` to make sure that the correct
GCing happens, and the migration occurs. One of the new tests
(additionally) acts as a regression test for this situation.