Fix large read (#343)
The current value we set for `MAX_KERNEL_BUF_SIZE` is not enough for
running node (version v22.19.0), which read ~4MB data from a file.
Instead of keeping updating this value (also 4MB happens to be the limit
our snp kernel allow to allocate), this PR tries to read data by chunks.