Fix page manager (#456)
This PR fixes the remaining issue in page manager mentioned in #452,
along with some other minor issues.

1. previously, `resize_mapping` called `insert_mapping` with
`fixed_address` to be true when it grew an existing mapping. However,
because the page manager is unaware of the memory allocated to the
global allocator, the requested memory may overlap with it and thus
corrupt it. Now it sets `fixed_address` to false so that it can detect
that the requested memory is unavailable when `insert_mapping` returns a
different memory address.
2. Propagate the error from platform to upper layers.
3. Config `TASK_ADDR_MAX` to 0xFFFF_F000 for x86 `LinuxUserland` because
running x86 programs on x64 platforms has a different address limit for
user space.
4. Update CI so that now it can test `litebox_shim_linux` on Windows
platform.
5. Fix `block_or_maybe_timeout`: according to
https://learn.microsoft.com/en-us/windows/win32/api/synchapi/nf-synchapi-waitonaddress,
`if the operation times out, GetLastError returns ERROR_TIMEOUT` not
`WAIT_TIMEOUT`.