Fixup lints in shim, fix a couple of bugs (#471)
Some of the allow declarations within the Linux shim were added _very_
early in its creation and should have been removed a long time ago. This
PR removes them and fixes up all the code that triggers warning when
those `allow` declarations are removed.

Interestingly, doing this surfaced a few bugs that were dormant in the
shim (all of these are now fixed):
- `recvfrom` would write beyond the length it was given
- `clock_getres` would silently drop errors and return "success"
- `SETFL` on pipes would silently drop errors and return "success"

The latter two issues are fairly minor, but the `recvfrom` bug is
particularly interesting, specifically:
- previously, it would entirely ignore the `len` parameter passed, so
the guest requesting for a buffer smaller than 4096 bytes could easily
get trampled. This would have been caught by the lint.
- assuming the user sends a `TRUNC` then we should actually copy over
the _real_ size, not the size specified via the output at `TRUNC`, so we
need to make sure we are taking the correct len. Once the lint-side
issue was fixed, this was caught by one of our tests.

So yeah, overall, lints + tests are being helpful, as long as we keep
them enabled :)

This PR is easiest to review by looking at each commit individually.