Clean up `SyscallRequest::try_from_raw` (#546)
This PR is a step towards #537. Rather than doing the entire thing in
one go, it is helpful to move in smaller steps.

This PR importantly introduces the comments reminding folks not to
increase the complexity of the parsing within
`SyscallRequest::try_from_raw`.

It also simplifies some of the existing code within it, by introducing a
`:?` combinator that simplifies a lot of repetition that existed.

Also, some of the hardcoded bits that existed wrt `TimeParam`, it
switches out to some slightly more generalized combinators; this _on its
face_ might seem like it is adding minor redundancy, but it is in fact
generalizing things out a little, mostly as a step towards eliminating
the dependency there. I've added comments to tell people not to use this
in any new places.

The intention (i.e., what this is moving towards) is to have the
flattened dispatch be possible while still having all the common stuff
be handled conveniently rather than every function needing to do its own
reinterpretation of inputs. However, this is a (small) step towards that
approach, and should at least remind us not to over-rely on parsing
here.